Move DataUpdateCoordinator to coordinator module in kraken
